pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "scrape-rethdb-data"
version = "0.1.0"
description = "High-performance Uniswap pool data collection from Reth database"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT OR Apache-2.0" }
keywords = ["ethereum", "uniswap", "reth", "database"]
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "eth-hash[pycryptodome]>=0.7.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "web3>=7.14.0",
]

[tool.maturin]
features = ["python"]
module-name = "scrape_rethdb_data"
